Update chapter-1-2.xhtml
As a result of a question, I looked something up in lint, and noticed that t-064 refers to z3998:morpheme, which isn't mentioned anywhere in the manual, which led me to see if it was used in the corpus, which led me to here.

For this particular issue, I'm doing this is a PR rather than just "fixing" it because you're the producer here, and presumably you know the manual better than the rest of us. :) But the examples in 8.2.8.1 would indicate this should be phoneme, not morpheme. Is that correct or no?

Then, should we add something to the manual to discuss when to use morpheme as opposed to phoneme/grapheme, and maybe an example or two? In looking at the existing uses in the corpus, several look odd to me, at least as I understand the definition of morpheme. In addition to this one, e.g., it's used for a word ("strike") in _Salammbo_, for a participle in _The Coming Race_, the words "through," "a," and "I" in Poe's _Short Fiction_, and so forth. Whether they are valid or not, the manual could help clarify what is valid use and what isn't.